test: extract shared editor test helpers into spec/js/editor_helpers.js
